+++ release_notes.md
@@ -1,6 +1,8 @@
+## 0.2.0
+
 ### Changed
 
-- Rename parameterless From method `FromDefaults` (#7)
+- Rename parameterless `From` method to `FromDefaults` (#7)
 
 ### New
 
@@ -9,7 +11,7 @@
 
 ### Additional Changes
 
-- Using AppVeyor for continuous build and test
+- Use AppVeyor for continuous build and test (#3)
 
 ## 0.1.0